+++ sdk/core/azure-core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.7.0-beta.1 (Unreleased)
 
+- Added `TokenRefreshOptions()` to `TokenCredential`, with a default token refresh offset of 2 minutes, and a default token refresh retry timeout of 30 seconds.
 
 ## 1.6.0 (2020-07-02)